move spatial view only image methods for image picking to spatial view crate
