+++ crates/viewer/re_data_ui/src/lib.rs
@@ -23,12 +23,7 @@ mod tensor;
 
 pub mod item_ui;
 
-pub use crate::{
-    image::{
-        show_zoomed_image_region, show_zoomed_image_region_area_outline, TextureInteractionId,
-    },
-    tensor::tensor_summary_ui_grid_contents,
-};
+pub use crate::tensor::tensor_summary_ui_grid_contents;
 pub use component::ComponentPathLatestAtResults;
 pub use component_ui_registry::{add_to_registry, register_component_uis};
