+++ crates/viewer/re_renderer/src/video/decoder/mod.rs
@@ -32,7 +32,10 @@ fn alloc_video_frame_texture(
             sample_count: 1,
             dimension: wgpu::TextureDimension::D2,
             format: wgpu::TextureFormat::Rgba8Unorm,
+            // Needs [`wgpu::TextureUsages::RENDER_ATTACHMENT`], otherwise copy of external textures will fail.
+            // Adding [`wgpu::TextureUsages::COPY_SRC`] so we can read back pixels on demand.
             usage: wgpu::TextureUsages::COPY_DST
+                | wgpu::TextureUsages::COPY_SRC
                 | wgpu::TextureUsages::TEXTURE_BINDING
                 | wgpu::TextureUsages::RENDER_ATTACHMENT,
         },
