+++ crates/viewer/re_space_view_spatial/src/lib.rs
@@ -15,6 +15,7 @@ mod mesh_loader;
 mod pickable_textured_rect;
 mod picking;
 mod picking_ui;
+mod picking_ui_pixel;
 mod proc_mesh;
 mod scene_bounding_boxes;
 mod space_camera_3d;
